% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfToSparse.R
\name{dfToSparse}
\alias{dfToSparse}
\title{Convert data frame to sparse matrix}
\usage{
dfToSparse(x, transpose = FALSE)
}
\arguments{
\item{x}{A data frame to convert into a sparse matrix.  Only character and
factor types are converted through one-hot encoding.}

\item{transpose}{A boolean determining if you should query by columns instead of
the default of rows (only for sparse matricies).}
}
\value{
A sparse matrix with one-hot encoded columns, numeric and integer
  columsn are preserved in this sparse matrix.  Column names are returned.
}
\description{
\code{dfToSparse} performs one-hot encoding to characters and factor types
within a data frame.  The encoding is returned as a sparse matrix.
}
\examples{
x <- matrix(runif(20),10,2)
neighbors <- dfToSparse(x_df) 
}
