% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_OverlappingCrowns2_function.R
\name{filter_OverlappingCrowns2}
\alias{filter_OverlappingCrowns2}
\title{Remove crown area polygons based on spatial overlap with other polygons}
\usage{
filter_OverlappingCrowns2(crowns, overlap.threshold = 0.9)
}
\arguments{
\item{crowns}{SpatialPolygonsDataFrame of crown projection areas, derived from make_CrownPolygons function}

\item{overlap.threshold}{Value between 0 and 1 specifying the relative overlap of a pair of crown projection areas (intersection area of two crowns divided by area of a single crown). If the relative overlap is for both crowns larger than this threshold, the crown polygon with the lower tree height value is removed.}
}
\value{
SpatialPolygonsDataFrame with each feature representing the crown projection area of one tree and columns containing various geometric attributes
}
\description{
The AMS3D algorithm can lead to clustering artefacts such that a single tree crown may be
assigned to more than one cluster ID and the derived enclosing crown polygons may show very
high spatial overlap. This function enables cleaning of the crown polygon dataset by removing
redundant polygons which most likely represent just one single tree. Of several
overlapping polygons, the polygon with the largest tree height value is kept, while the other
polygons are removed.
The algorithm involves pairwise comparisons of all polygons which are handed over to the
function. Hence, it is recommended to apply it on small tiles (e.g., 1 ha), as processing
time may become very slow for large areas.
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
\keyword{area}
\keyword{artefact}
\keyword{clean}
\keyword{cluster}
\keyword{crown}
\keyword{discard}
\keyword{double}
\keyword{drop}
\keyword{error}
\keyword{filter}
\keyword{overlap}
\keyword{polygon}
\keyword{projection}
\keyword{remove}
\keyword{tree}
