% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_CrownPolygons_function.R
\name{make_CrownPolygons}
\alias{make_CrownPolygons}
\title{Fit polygon hulls around tree crowns in a clustered point cloud to derive crown projection areas}
\usage{
make_CrownPolygons(pc.dt, type = "convexhull", N.min = 20,
  Ext.min = 2, Ext.max = 50, proj4string = CRS(as.character(NA)))
}
\arguments{
\item{pc.dt}{Point cloud in data.table format containing columns X, Y, Z and ID}

\item{type}{Shape of the desired polygons ("convexhull", "ellipse" or "circle")}

\item{N.min}{Minimum number of points in a cluster to be considered as a tree crown}

\item{Ext.min}{Minimum horizontal extent (in meters, in X- and Y-direction) of a cluster to be considered as a tree crown}

\item{Ext.max}{Maximum horizontal extent (in meters, in X- and Y-direction) of a cluster to be considered as a tree crown}

\item{proj4string}{Projection string of class CRS-class}
}
\value{
SpatialPolygonsDataFrame with each feature representing the crown projection area of one tree and columns containing various geometric attributes
}
\description{
The function creates crown projection area polygons from a clustered point cloud of a forest stand.
The input point cloud needs to have a column containing tree ID / cluster ID of each point.
There are different options for the shape of the polygons.
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
\keyword{area}
\keyword{circle}
\keyword{cloud}
\keyword{cluster}
\keyword{convex}
\keyword{crown}
\keyword{ellipse}
\keyword{hull}
\keyword{perimeter}
\keyword{point}
\keyword{polygons}
\keyword{projection}
\keyword{tree}
