% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_CrownsStems_function.R
\name{match_CrownsStems}
\alias{match_CrownsStems}
\title{Match crown projection polygons with tree stem positions}
\usage{
match_CrownsStems(crowns.spdf, stems.spdf, DBH.min = 0.05)
}
\arguments{
\item{crowns.spdf}{SpatialPolygonsDataFrame of tree crown projection areas}

\item{stems.spdf}{SpatialPointsDataFrame of tree stem foot positions (required columns: X, Y, DBH, Species, TreeID)}

\item{DBH.min}{Minimum stem diameter to be considered}
}
\value{
SpatialPolygonsDataFrame of tree crown projection areas with additional columns containing the attributes of the assigned stem for each crown
}
\description{
The function links remote sensing derived crown projection polygons with inventory derived tree stems.
The algorithm sorts all polygons by decreasing area size. Then, starting with the largest crown it checks all
stem positions that fall inside the crown projection area and selects the stem with the largest diameter
and assigns it to the crown. The stem is removed from the set of available stems and the same procedure
is conducted for the next smaller crown and further on until the smallest crown.
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
\keyword{DBH}
\keyword{area}
\keyword{crown}
\keyword{diameter}
\keyword{foot}
\keyword{ground}
\keyword{inventory}
\keyword{link}
\keyword{match}
\keyword{point}
\keyword{polygon}
\keyword{position}
\keyword{projection}
\keyword{stem}
\keyword{tree}
\keyword{truth}
