\name{meanShiftR-package}
\alias{meanShiftR-package}
\alias{meanShiftR}
\docType{package}
\title{
\packageTitle{meanShiftR}
}
\description{
\packageDescription{meanShiftR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{meanShiftR}
\packageIndices{meanShiftR}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{meanShiftR}

Maintainer: \packageMaintainer{meanShiftR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
