% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_BufferedPointCloud_function.R
\name{split_BufferedPointCloud}
\alias{split_BufferedPointCloud}
\title{Split point cloud into smaller point clouds with buffer area around}
\usage{
split_BufferedPointCloud(pc.dt, plot.width, buffer.width)
}
\arguments{
\item{pc.dt}{Point cloud in data.table format containing columns X, Y and Z}

\item{plot.width}{Width of the core area in meters}

\item{buffer.width}{Width of the buffer around the core area in meters}
}
\value{
List of sub point clouds and buffer and core points labeled in boolean column "Buffer"
}
\description{
The function splits one large point cloud into several smaller point clouds.
It allows to specify a buffer width around the core area where points are included.
}
\author{
Nikolai Knapp, nikolai.knapp@ufz.de
}
\keyword{area}
\keyword{buffer}
\keyword{cloud}
\keyword{parallel}
\keyword{plot}
\keyword{point}
\keyword{split}
\keyword{subset}
